agent/requirements.txt
anthropic>=0.52.0
python-dotenv>=1.0.0
requests>=2.32.0
